update to upstream rev. e575d02196b0013c25f8064e4dbe3fc2a0ef72d0

+++ meta/classes/kernel.bbclass
@@ -452,10 +452,7 @@ kernel_do_install() {
 		rm -f "${D}${nonarch_base_libdir}/modules/${KERNEL_VERSION}/build"
 		rm -f "${D}${nonarch_base_libdir}/modules/${KERNEL_VERSION}/source"
 		# Remove empty module directories to prevent QA issues
-		# MobiAqua: check if directory exists first
-		if [ -d "${D}${nonarch_base_libdir}/modules/${KERNEL_VERSION}/kernel" ]; then
-			find "${D}${nonarch_base_libdir}/modules/${KERNEL_VERSION}/kernel" -type d -empty -delete
-		fi
+		[ -d "${D}${nonarch_base_libdir}/modules/${KERNEL_VERSION}/kernel" ] && find "${D}${nonarch_base_libdir}/modules/${KERNEL_VERSION}/kernel" -type d -empty -delete
 	else
 		bbnote "no modules to install"
 	fi
